- lint:cast
git-svn-id: https://scm.gforge.inria.fr/svn/alignapi/trunk@1275 a7222439-140d-0410-8486-c7398dd066c2

+++ src/fr/inrialpes/exmo/align/util/GroupOutput.java
@@ -248,7 +248,7 @@ public class GroupOutput {
 	// Unload the ontologies.
 	OntologyFactory.clear();
 
-	return (double)result/(double)tests.length;
+	return result/(double)tests.length;
     }
 
     public Evaluator eval( String alignName1, String alignName2 ) throws AlignmentException {
